chore: try remove dp model
