+++ deepmd/dpmodel/atomic_model/polar_atomic_model.py
@@ -1,0 +1,9 @@
+from .dp_atomic_model import (
+    DPAtomicModel,
+)
+
+class DPPolarAtomicModel(DPAtomicModel):
+    
+    def apply_out_stat(self, ret, atype):
+         # TODO: migrate bias
+        pass
