+++ deepmd/pt/model/model/polar_model.py
@@ -6,12 +6,19 @@ from typing import (
 
 import torch
 
-from .dp_model import (
-    DPModel,
+
+from .make_model import (
+    make_model,
+)
+from deepmd.pt.model.atomic_model import(
+    DPPolarAtomicModel
 )
 
-
-class PolarModel(DPModel):
+from deepmd.pt.model.model.model import (
+    BaseModel,
+)
+@BaseModel.register("standard")
+class PolarModel(make_model(DPPolarAtomicModel)):
     model_type = "polar"
 
     def __init__(
@@ -49,6 +56,31 @@ class PolarModel(DPModel):
             model_predict["updated_coord"] += coord
         return model_predict
 
+    @classmethod
+    def update_sel(cls, global_jdata: dict, local_jdata: dict):
+        """Update the selection and perform neighbor statistics.
+
+        Parameters
+        ----------
+        global_jdata : dict
+            The global data, containing the training section
+        local_jdata : dict
+            The local data refer to the current class
+        """
+        local_jdata_cpy = local_jdata.copy()
+        local_jdata_cpy["descriptor"] = cls.get_descriptor().update_sel(
+            global_jdata, local_jdata["descriptor"]
+        )
+        return local_jdata_cpy
+
+    def get_fitting_net(self):
+        """Get the fitting network."""
+        return self.atomic_model.fitting_net
+
+    def get_descriptor(self):
+        """Get the descriptor."""
+        return self.atomic_model.descriptor
+    
     @torch.jit.export
     def forward_lower(
         self,
